+++ app/models/assignment.rb
@@ -27,7 +27,8 @@ class Assignment < ApplicationRecord
   belongs_to :user
   has_many :selected_documents, through: :student_works
   has_many :student_works, dependent: :destroy
-  has_many :feedback_items, as: :feedbackable
+
+  has_one :assignment_summary, dependent: :destroy
 
   # Grade levels for assignment form
   GRADE_LEVELS = [ "5", "6", "7", "8", "9", "10", "11", "12", "university" ].freeze